+++ wemake_python_styleguide/types.py
@@ -114,3 +114,4 @@ class ConfigurationOptions(Protocol):
     max_try_body_length: int
     max_module_expressions: int
     max_function_expressions: int
+    max_asserts: int